Move routes to ./routes.js & convert last few definitions to app_route
